Add 'page.js' and reorganize functions
